chore: cleanup, refactor
