+++ protocol-units/da-sequencer/node/src/tests/mod.rs
@@ -2,6 +2,7 @@ use crate::batch::*;
 use crate::run;
 use crate::server::{run_server, GrpcRequests};
 use crate::tests::mock::{CelestiaMock, StorageMock};
+use crate::tests::whitelist::make_test_whitelist;
 use crate::whitelist::Whitelist;
 use ed25519_dalek::Signature;
 use movement_da_sequencer_client::{sign_batch, DaSequencerClient};
@@ -17,86 +18,7 @@ use tokio::time::Duration;
 use tracing_subscriber;
 
 pub mod mock;
-
-fn make_test_whitelist(keys: Vec<ed25519_dalek::VerifyingKey>) -> Whitelist {
-	Whitelist::from_keys(keys)
-}
-
-#[serial]
-#[tokio::test]
-async fn test_sign_and_validate_batch_passes_with_whitelisted_signer() {
-	let _ = tracing_subscriber::fmt()
-		.with_max_level(tracing::Level::INFO)
-		.with_test_writer()
-		.try_init();
-
-	let config = DaSequencerConfig::default();
-	let signing_key = config.signing_key;
-	let verifying_key = signing_key.verifying_key();
-
-	let whitelist = Whitelist::from_keys(vec![verifying_key]);
-
-	let txs = FullNodeTxs(vec![
-		Transaction::new(b"hello".to_vec(), 0, 1),
-		Transaction::new(b"world".to_vec(), 0, 2),
-	]);
-
-	let batch_bytes = bcs::to_bytes(&txs).expect("Serialization failed");
-	let signature = sign_batch(&batch_bytes, &signing_key);
-	let serialized =
-		serialize_full_node_batch(verifying_key, signature.clone(), batch_bytes.clone());
-
-	let (deserialized_key, deserialized_sig, deserialized_data) =
-		deserialize_full_node_batch(serialized).expect("Deserialization failed");
-
-	let raw_batch = DaBatch {
-		data: RawData { data: deserialized_data },
-		signature: deserialized_sig,
-		signer: deserialized_key,
-		timestamp: chrono::Utc::now().timestamp_micros() as u64,
-	};
-
-	let validated = validate_batch(raw_batch, &whitelist).expect("Batch should validate");
-	assert_eq!(validated.data.0, txs.0);
-}
-
-#[serial]
-#[tokio::test]
-async fn test_sign_and_validate_batch_fails_with_non_whitelisted_signer() {
-	let _ = tracing_subscriber::fmt()
-		.with_max_level(tracing::Level::INFO)
-		.with_test_writer()
-		.try_init();
-
-	let whitelist = Whitelist::from_keys(vec![]); // empty whitelist
-
-	let config = DaSequencerConfig::default();
-	let signing_key = config.signing_key;
-	let verifying_key = signing_key.verifying_key();
-
-	let txs = FullNodeTxs(vec![
-		Transaction::new(b"hello".to_vec(), 0, 1),
-		Transaction::new(b"world".to_vec(), 0, 2),
-	]);
-
-	let batch_bytes = bcs::to_bytes(&txs).expect("Serialization failed");
-	let signature = sign_batch(&batch_bytes, &signing_key);
-	let serialized =
-		serialize_full_node_batch(verifying_key, signature.clone(), batch_bytes.clone());
-
-	let (deserialized_key, deserialized_sig, deserialized_data) =
-		deserialize_full_node_batch(serialized).expect("Deserialization failed");
-
-	let raw_batch = DaBatch {
-		data: RawData { data: deserialized_data },
-		signature: deserialized_sig,
-		signer: deserialized_key,
-		timestamp: chrono::Utc::now().timestamp_micros() as u64,
-	};
-
-	let result = validate_batch(raw_batch, &whitelist);
-	assert!(matches!(result, Err(crate::error::DaSequencerError::InvalidSigner)));
-}
+pub mod whitelist;
 
 #[tokio::test]
 #[serial]
